+++ examples/my_code/scripts/backup.exs
@@ -1,10 +1,10 @@
 defmodule Backup do
-  @inputs "{config,lib,test}/**/*"
+  @inputs "{priv,config,lib,test}/**/*"
   @backup "scripts/backup.bin"
 
   def run([]) do
     @inputs
-    |> Path.wildcard()
+    |> Path.wildcard(match_dot: true)
     |> Enum.reject(&File.dir?/1)
     |> Enum.into(%{}, fn path -> {path, File.read!(path)} end)
     |> backup()