+++ lib/recode/task/unnecessary_if_unless.ex
@@ -19,8 +19,6 @@ defmodule Recode.Task.UnnecessaryIfUnless do
 
   use Recode.Task, corrector: true, category: :readability
 
-  alias Recode.Issue
-  alias Recode.Task.UnnecessaryIfUnless
   alias Rewrite.Source
   alias Sourceror.Zipper
 
@@ -34,13 +32,7 @@ defmodule Recode.Task.UnnecessaryIfUnless do
         remove_unecessary_if_unless(zipper, issues, opts[:autocorrect])
       end)
 
-    case opts[:autocorrect] do
-      true ->
-        Source.update(source, UnnecessaryIfUnless, :quoted, Zipper.root(zipper))
-
-      false ->
-        Source.add_issues(source, issues)
-    end
+    update_source(source, opts, quoted: zipper, issues: issues)
   end
 
   defp remove_unecessary_if_unless(
@@ -70,7 +62,7 @@ defmodule Recode.Task.UnnecessaryIfUnless do
       case extract(body, conditional) do
         {:ok, _expr} ->
           message = "Avoid unnecessary `if` and `unless`"
-          issue = Issue.new(UnnecessaryIfUnless, message, meta)
+          issue = new_issue(message, meta)
           [issue | issues]
 
         :error ->